+++ pxemgr/ipxe_handlers.go
@@ -36,9 +36,9 @@ func (mgr *pxeManagerT) ipxeBootScript(w http.ResponseWriter, r *http.Request) {
 		_ = mgr.logger.Log("level", "info", "message", "adding 'console=ttyS0' to kernel args")
 	}
 
-	if mgr.coreosAutologin {
-		extraFlags += " coreos.autologin"
-		_ = mgr.logger.Log("level", "info", "message", "adding coreos.autologin to kernel args")
+	if mgr.flatcarAutologin {
+		extraFlags += " flatcar.autologin"
+		_ = mgr.logger.Log("level", "info", "message", "adding flatcar.autologin to kernel args")
 	}
 
 	if mgr.systemdShell {
@@ -47,7 +47,7 @@ func (mgr *pxeManagerT) ipxeBootScript(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// for ignition we use only 1phase installation without mayu-infopusher
-	kernel := fmt.Sprintf("kernel %s/images/vmlinuz coreos.first_boot=1 initrd=initrd.cpio.gz coreos.config.url=%s?uuid=${uuid}&serial=${serial} systemd.journald.max_level_console=debug verbose log_buf_len=10M "+extraFlags+"\n", mgr.pxeURL(), mgr.ignitionURL())
+	kernel := fmt.Sprintf("kernel %s/images/vmlinuz flatcar.first_boot=1 initrd=initrd.cpio.gz flatcar.config.url=%s?uuid=${uuid}&serial=${serial} systemd.journald.max_level_console=debug verbose log_buf_len=10M "+extraFlags+"\n", mgr.pxeURL(), mgr.ignitionURL())
 	initrd := fmt.Sprintf("initrd %s/images/initrd.cpio.gz\n", mgr.pxeURL())
 	// console=ttyS0,115200n8
 	buffer.WriteString("#!ipxe\n")
@@ -156,13 +156,13 @@ func (mgr *pxeManagerT) imagesHandler(w http.ResponseWriter, r *http.Request) {
 	var img *os.File
 	var err error
 
-	coreOSversion := mgr.config.DefaultFlatcarVersion
-	_ = mgr.logger.Log("level", "info", "message", fmt.Sprintf("sending Container Linux %s image", coreOSversion))
+	flatcarVersion := mgr.config.DefaultFlatcarVersion
+	_ = mgr.logger.Log("level", "info", "message", fmt.Sprintf("sending Container Linux %s image", flatcarVersion))
 
 	if strings.HasSuffix(r.URL.Path, "/vmlinuz") {
-		img, err = mgr.pxeKernelImage(coreOSversion)
+		img, err = mgr.pxeKernelImage(flatcarVersion)
 	} else if strings.HasSuffix(r.URL.Path, "/initrd.cpio.gz") {
-		img, err = mgr.pxeInitRD(coreOSversion)
+		img, err = mgr.pxeInitRD(flatcarVersion)
 	} else {
 		panic(fmt.Sprintf("no handler provided for invalid URL path '%s'", r.URL.Path))
 	}