+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Switch all legacy `coreos` flags to `flatcar` (drops CoreOS support).
+
 ## [1.2.0] - 2021-06-09
 
 ### Added